rename: Schedule 컴포넌트, 컨텍스트, 훅 추후 배포를 위한 폴더 분리
