+++ airbnb/src/lib/Schedule/index.ts
@@ -1,0 +1,1 @@
+export { Schedule } from './Components/Schedule';
